--- src/main/java/convertibleUnits/unitConversion/UnitConversionInterface.java
@@ -1,11 +1,0 @@
-package convertibleUnits.unitConversion;
-
-/**
- * TODO: finish javadocs
- */
-public interface UnitConversionInterface {
-    /**
-     * TODO: finish this javadoc
-     */
-    double convert(double value);
-}
