--- src/main/java/timeConversionModel2/convertibleUnits/NullUnitConversionException.java
@@ -1,13 +1,0 @@
-package timeConversionModel2.convertibleUnits;
-
-/**
- * TODO: finish javadocs
- */
-public final class NullUnitConversionException extends Exception {
-    /**
-     * TODO: finish this javadoc
-     */
-    public NullUnitConversionException() {
-        super("Unit conversion not defined!");
-    }
-}
