--- src/main/java/timeConversionModel2/mathOperations/SingleVariableOperation.java
@@ -1,6 +1,0 @@
-package timeConversionModel2.mathOperations;
-
-//TODO: add javadocs
-interface SingleVariableOperation {
-    double calculate(double variable);
-}
