--- src/main/java/timeConversionModel2/referancableUnits/ReferencableUnit.java
@@ -1,4 +1,0 @@
-package timeConversionModel2.referancableUnits;
-
-//
-public abstract class ReferencableUnit {}
