+++ src/main/java/unitConversion/convertibleUnits/UnitConversionTable.java
@@ -1,4 +1,4 @@
-package timeConversionModel2.convertibleUnits;
+package unitConversion.convertibleUnits;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -6,7 +6,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Map;
 import java.util.HashMap;
 
-import timeConversionModel2.mathOperations.SingleCoefficientOperation;
+import unitConversion.mathOperations.SingleCoefficientOperation;
 
 //TODO: add javadocs
 public abstract class UnitConversionTable<T extends ConvertibleUnit> {
