+++ src/main/java/unitConversion/mathOperations/SingleVariableOperation.java
@@ -1,0 +1,6 @@
+package unitConversion.mathOperations;
+
+//TODO: add javadocs
+public interface SingleVariableOperation {
+    double calculate(double variable);
+}
