+++ src/main/java/unitConversion/referancableUnits/ReferencableUnit.java
@@ -1,0 +1,4 @@
+package unitConversion.referancableUnits;
+
+//A class for the unit to be used within UnitReferenceTable
+public abstract class ReferencableUnit {}
